XBARA version based on FSL
see examples/eFlexPwmFault
Does not work. As soon as fault management is activated with
Timer::setupFaults(), the FAULTx input pins are seen high, which is
confirmed by the FFPIN bits in the FSTS register, yet the FFLAG bits
remain high. the low state.
This does not change even if the FAULTx inputs are routed high or low
with XBAR.
The only thing that disables the error on startup is the
setFaultLevel(true) configuration but in this case it is not possible to
make the timer go into fault.